+++ src/codemodder/codemods/base_codemod.py
@@ -108,10 +108,10 @@ class BaseCodemod(metaclass=ABCMeta):
 
     @cached_property
     def description(self) -> str:
-        if self._metadata.description == None:
+        if self._metadata.description is None:
             doc_path = self.docs_module / f"{self.origin}_python_{self.name}.md"
             return doc_path.read_text()
-        return self._metadata.description  # type: ignore
+        return self._metadata.description
 
     @property
     def review_guidance(self):
