+++ src/core_codemods/__init__.py
@@ -50,7 +50,7 @@ from .fix_empty_sequence_comparison import FixEmptySequenceComparison
 from .remove_assertion_in_pytest_raises import RemoveAssertionInPytestRaises
 from .fix_assert_tuple import FixAssertTuple
 
-from .sonar_numpy_nan_equality import SonarNumpyNanEquality
+from .sonar.sonar_numpy_nan_equality import SonarNumpyNanEquality
 
 registry = CodemodCollection(
     origin="pixee",
